Set local ip as gateway instead of charon ip
Setting the charon ip on routes makes it impossible to register
cross IP version tunnels with pfroute. Since it doesn't make sense
to use this ip as the default gateway in the pfroute implementation
anyway (all traffic is sent into tun devices), we're setting
the source ip as default gateway, since this IP is guaranteed to be
of the same IP version as the packets in the IPsec tunnel.