wip: vici: Send all queued messages during shutdown
This ensures that e.g. ike/child-updown messages are sent that were
queued but couldn't be sent (even the job to enable to on_write() callback
requires a worker thread that's not around anymore during shutdown).

wip: Not sure if sending the messages blocking could be a problem.
Maybe it could cause a lockup if lots of messages are sent that a client
is not reading (not actually sure what the conditions must be for the
write to block). So we could perhaps do some best-effort thing and
just send them non-blocking and just abort when the write would block..

 #3602