wip: vici: Expose ike-update event
wip: Not sure if children are required (they might if e.g. routes have to be
(re-)installed via a new tunnel interface).  The question is if it would
be better to add a child-update event too. But that might cause a lot of
events, so this approach could be simpler to handle such changes in clients.

The syntax for passing the new address/port might also not be ideal.
Perhaps passing new-local-host/port and new-remote-host/port (only the
ones set that changed, also see below) might be more usable.

On a similar note it might be useful to only send this event once if
both local and remote host change (e.g. a switch from IPv4 to IPv6), as
currently there will be one for the local and then the remote address.
We could e.g. change the ike_updown() bus/listener method to take two
addresses and only set the one that changed and in the vici message
add new-local-host/port and new-remote-host/port depending on which ones
are actually passed.